FIXED BUG WITH ARCHIVE VERSIONS
The behavior for finding archived versions,
when the user passes the -d/--date arguments
for a date different than the current date,
was incorrect and didn't work. The os.path.join
on line 255 was passing in the note_filename
variable as the last argument. because this
was already a full path, we were not seeing
the archived directory getting built correctly.